+++ index.js
@@ -1,5 +1,9 @@
 import {createElement, Component, createContext} from 'react'
 
+/* START DEVELOPMENT BUILD ONLY */
+import devTool from './dev-tool'
+/* END DEVELOPMENT BUILD ONLY */
+
 import factory from './factory'
 
-export default factory(createElement, Component, createContext)
+export default factory(devTool, createElement, Component, createContext)